% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_nodes.R
\name{adaptive_nodes}
\alias{adaptive_nodes}
\title{Find a set of nodes to cover a dataset}
\usage{
adaptive_nodes(data, n_nodes = 10, max.it = 100)
}
\arguments{
\item{data}{A matrix of data points. Each row should be a data point.}

\item{n_nodes}{The number of nodes to cover the dataset.}

\item{max.it}{The number of iterations to use when finding the best set of nodes.}
}
\value{
A list giving the nodes and the objective function trajectory.
}
\description{
Find a set of nodes to cover a dataset
}
